server/requirements.txt
flask
requests
openmeteo-requests
python-dotenv
pytz
pandas
numpy
torch
scikit-learn
influxdb
paho-mqtt